mejoras de la web

+++ services/necesidades/src/main/java/com/donaton/necesidades/dto/NecesidadEvent.java
@@ -11,7 +11,7 @@ public class NecesidadEvent {
 
     private Long id;
     private String titulo;
-    private Integer cantidadRequerida;
+    private Double cantidadRequerida;
     private String ubicacion;
     private String estado;
 }
